Reorganized EditUserPage components along with amended design.
