+++ frontend/src/components/user/UserControlMenu.tsx
@@ -94,6 +94,7 @@ export const UserControlMenu: FC<UserControlProps> = ({
           userId={user.id}
           openModal={openModal}
           onClose={handleCloseModal}
+          onSubmit={() => {}}  // This is used at page/EditUserPage component
         />
       </Box>
     </Menu>
