+++ frontend/src/components/user/UserPasswordFormModal.test.tsx
@@ -27,6 +27,9 @@ test("should render a component with essential props", function () {
         onClose={() => {
           /* dummy */
         }}
+        onSubmit={() => {
+          /* dummy */
+        }}
       />,
       {
         wrapper: TestWrapper,
